requirements.txt
flask>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
spacy>=3.7.0
# Optional: for sentence-transformer embeddings (better NLP accuracy)
# sentence-transformers>=2.2.0
pypdf>=3.17.0
python-docx>=1.1.0
